+++ test/cypress/integration/addFormToPage-error-message.cy.js
@@ -8,7 +8,7 @@ describe('addFormToPage Error Message Test', () => {
     cy.get('.lhc-form-title').contains('A questionnaire for testing code that requires a FHIR context 1');
     // has an error message
     // This also tests that the url parameter gets URL-encoded.
-    cy.get("#loadMsg").contains("Unable to load ValueSet from https://lforms-fhir.nlm.nih.gov/baseDstu3/ValueSet/$expand?url=http%3A%2F%2Fhl7.org%2Ffhir%2FValueSet%2Fyesnodontknow-invalid");
+    cy.get("#loadMsg").contains("Unable to load ValueSet");
   });
 
   it('show an error when a valueset cannot be loaded because of a wrong fhir context', () => {